+++ packages/cloud-cognitive/src/components/DataSpreadsheet/DataSpreadsheetHeader.js
@@ -12,8 +12,11 @@ import { px } from '@carbon/layout';
 import { pkg } from '../../settings';
 import { usePreviousValue } from '../../global/js/hooks';
 import { checkActiveHeaderCell } from './utils/checkActiveHeaderCell';
+import { checkSelectedHeaderCell } from './utils/checkSelectedHeaderCell';
 import { handleHeaderCellSelection } from './utils/handleHeaderCellSelection';
 import { selectAllCells } from './utils/selectAllCells';
+import { getSpreadsheetWidth } from './utils/getSpreadsheetWidth';
+import { useSpreadsheetMouseMove } from './hooks';
 
 const blockClass = `${pkg.prefix}--data-spreadsheet`;
 
@@ -32,11 +35,16 @@ export const DataSpreadsheetHeader = forwardRef(
       setSelectionAreas,
       setSelectionAreaData,
       rows,
+      totalVisibleColumns,
       updateActiveCellCoordinates,
+      setHeaderCellHoldActive,
+      headerCellHoldActive,
     },
     ref
   ) => {
     const [scrollBarSizeValue, setScrollBarSizeValue] = useState(0);
+    const [selectedHeaderReorderActive, setSelectedHeaderReorderActive] =
+      useState(false);
     const previousState = usePreviousValue({ cellSize });
     useEffect(() => {
       if (previousState?.cellSize !== cellSize) {
@@ -81,6 +89,44 @@ export const DataSpreadsheetHeader = forwardRef(
       });
     };
 
+    const handleHeaderMouseDown = (index) => {
+      return (event) => {
+        setSelectedHeaderReorderActive(true);
+        const clickXPosition = event.clientX;
+        const headerButtonCoords = event.target.getBoundingClientRect();
+        const offsetXValue = clickXPosition - headerButtonCoords.left;
+        const bodyContainer = document.querySelector(
+          `.${blockClass}__list--container`
+        ).firstElementChild;
+        const selectionAreaToClone = selectionAreas.filter(
+          (item) => item?.header?.index === index
+        );
+        const selectionAreaElement = ref.current.querySelector(
+          `[data-matcher-id="${selectionAreaToClone[0]?.matcher}"]`
+        );
+        const selectionAreaClonedElement = selectionAreaElement.cloneNode();
+        const reorderIndicatorLine = selectionAreaElement.cloneNode();
+        reorderIndicatorLine.className = `${blockClass}__reorder-indicator-line`;
+        reorderIndicatorLine.style.width = px(2);
+        selectionAreaClonedElement.classList.add(
+          `${blockClass}__selection-area--element-cloned`
+        );
+        selectionAreaClonedElement.setAttribute(
+          'data-clone-offset-x',
+          offsetXValue
+        );
+        selectionAreaClonedElement.setAttribute(
+          'data-column-index-original',
+          index
+        );
+        bodyContainer.appendChild(selectionAreaClonedElement);
+        bodyContainer.appendChild(reorderIndicatorLine);
+        setHeaderCellHoldActive(true);
+      };
+    };
+
+    useSpreadsheetMouseMove({ ref, headerCellHoldActive, defaultColumn });
+
     return (
       <div className={cx(`${blockClass}__header--container`)} role="rowgroup">
         {headerGroups.map((headerGroup, index) => (
@@ -89,18 +135,23 @@ export const DataSpreadsheetHeader = forwardRef(
             {...headerGroup.getHeaderGroupProps()}
             style={{
               ...headerGroup.getHeaderGroupProps().style,
-              width: px(
-                parseInt(headerGroup.getHeaderGroupProps().style.width) +
-                  scrollBarSizeValue
-              ),
+              width: getSpreadsheetWidth({
+                type: 'header',
+                headerGroup,
+                scrollBarSizeValue,
+                totalVisibleColumns,
+                defaultColumn,
+              }),
+              overflow: 'hidden',
             }}
             className={`${blockClass}__tr`}
           >
             {/* SELECT ALL BUTTON */}
             <div
               role="columnheader"
+              className={`${blockClass}__select-all-cell-container`}
               style={{
-                width: defaultColumn?.rowHeaderWidth - 4,
+                width: defaultColumn?.rowHeaderWidth,
                 height: defaultColumn?.rowHeight,
               }}
             >
@@ -126,37 +177,62 @@ export const DataSpreadsheetHeader = forwardRef(
               </button>
             </div>
             {/* COLUMN HEADER BUTTONS */}
-            {headerGroup.headers.map((column, index) => (
-              <div
-                key={`column_${index}`}
-                role="columnheader"
-                className={`${blockClass}__columnheader`}
-                {...column.getHeaderProps()}
-              >
-                <button
-                  id={`${blockClass}__cell--header--${index}`}
-                  data-row-index="header"
-                  data-column-index={index}
-                  tabIndex={-1}
-                  onClick={handleColumnHeaderClick(index)}
-                  style={{
-                    height: defaultColumn?.rowHeight,
-                  }}
-                  className={cx(
-                    `${blockClass}__th`,
-                    `${blockClass}--interactive-cell-element`,
-                    {
-                      [`${blockClass}__th--active-header`]:
-                        activeCellCoordinates?.column === index ||
-                        checkActiveHeaderCell(index, selectionAreas, 'column'),
-                    }
-                  )}
-                  type="button"
+            {headerGroup.headers.map((column, index) => {
+              const selectedHeader = checkSelectedHeaderCell(
+                index,
+                selectionAreas,
+                'column'
+              );
+              return (
+                <div
+                  key={`column_${index}`}
+                  role="columnheader"
+                  className={`${blockClass}__columnheader`}
+                  {...column.getHeaderProps()}
                 >
-                  {column.render('Header')}
-                </button>
-              </div>
-            ))}
+                  <button
+                    id={`${blockClass}__cell--header--${index}`}
+                    data-row-index="header"
+                    data-column-index={index}
+                    tabIndex={-1}
+                    onMouseDown={
+                      selectedHeader ? handleHeaderMouseDown(index) : null
+                    }
+                    onMouseUp={
+                      selectedHeader
+                        ? () => setSelectedHeaderReorderActive(false)
+                        : null
+                    }
+                    onClick={
+                      !selectedHeader ? handleColumnHeaderClick(index) : null
+                    }
+                    style={{
+                      height: defaultColumn?.rowHeight,
+                      width: column?.width || defaultColumn?.width,
+                    }}
+                    className={cx(
+                      `${blockClass}__th`,
+                      `${blockClass}--interactive-cell-element`,
+                      {
+                        [`${blockClass}__th--active-header`]:
+                          activeCellCoordinates?.column === index ||
+                          checkActiveHeaderCell(
+                            index,
+                            selectionAreas,
+                            'column'
+                          ),
+                        [`${blockClass}__th--selected-header`]: selectedHeader,
+                        [`${blockClass}__th--selected-header-reorder-active`]:
+                          selectedHeaderReorderActive,
+                      }
+                    )}
+                    type="button"
+                  >
+                    {column.render('Header')}
+                  </button>
+                </div>
+              );
+            })}
           </div>
         ))}
       </div>
@@ -193,6 +269,11 @@ DataSpreadsheetHeader.propTypes = {
   }),
 
   /**
+   * Whether or not a click/hold is active on a header cell
+   */
+  headerCellHoldActive: PropTypes.bool,
+
+  /**
    * Headers provided from useTable hook
    */
   headerGroups: PropTypes.arrayOf(PropTypes.object),
@@ -223,6 +304,11 @@ DataSpreadsheetHeader.propTypes = {
   setCurrentMatcher: PropTypes.func,
 
   /**
+   * Setter fn for header cell hold active value
+   */
+  setHeaderCellHoldActive: PropTypes.func,
+
+  /**
    * Setter fn for selectionAreaData state value
    */
   setSelectionAreaData: PropTypes.func,
@@ -231,6 +317,12 @@ DataSpreadsheetHeader.propTypes = {
    * Setter fn for selectionAreas value
    */
   setSelectionAreas: PropTypes.func,
+
+  /**
+   * The total number of columns to be initially visible, additional columns will be rendered and
+   * visible via horizontal scrollbar
+   */
+  totalVisibleColumns: PropTypes.number,
 
   /**
    * Function used to update the active cell coordinates