+++ packages/cloud-cognitive/src/components/DataSpreadsheet/hooks/index.js
@@ -9,3 +9,6 @@ export { useMoveActiveCell } from './useMoveActiveCell';
 export { useMultipleKeyTracking } from './useMultipleKeyTracking';
 export { useResetSpreadsheetFocus } from './useResetSpreadsheetFocus';
 export { useSpreadsheetOutsideClick } from './useSpreadsheetOutsideClick';
+export { useSpreadsheetEdit } from './useSpreadsheetEdit';
+export { useSpreadsheetMouseUp } from './useSpreadsheetMouseUp';
+export { useSpreadsheetMouseMove } from './useSpreadsheetMouseMove';