+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/DatagridHead.js
@@ -1,10 +1,10 @@
-/*
- * Licensed Materials - Property of IBM
- * 5724-Q36
- * (c) Copyright IBM Corp. 2020, 2021
- * US Government Users Restricted Rights - Use, duplication or disclosure
- * restricted by GSA ADP Schedule Contract with IBM Corp.
+/**
+ * Copyright IBM Corp. 2020, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
  */
+
 import React from 'react';
 import { DataTable } from 'carbon-components-react';
 
@@ -12,19 +12,12 @@ const { TableHead } = DataTable;
 
 const DatagridHead = (datagridState) => {
   const { headerGroups = [], headRef, HeaderRow } = datagridState;
-
   return (
     <TableHead>
-      {headerGroups.map((headerGroup, index) => (
+      {headerGroups.map((headerGroup) =>
         // doesn't support header grouping.
-        <div
-          {...headerGroup.getHeaderGroupProps()}
-          ref={headRef}
-          key={`header_${index}`}
-        >
-          {HeaderRow(datagridState)}
-        </div>
-      ))}
+        HeaderRow(datagridState, headRef, headerGroup)
+      )}
     </TableHead>
   );
 };