+++ packages/security/src/components/TrendingCard/TrendingCard-story.js
@@ -7,9 +7,10 @@ import React from 'react';
 
 import { components, getDocsParameters } from '../../../.storybook';
 import { TrendingCard } from '../..';
+import { InlineNotification, NotificationActionButton } from '../../';
 
 export default {
-  title: components(TrendingCard.name),
+  title: components(`${TrendingCard.name}#legacy`),
   component: TrendingCard,
   args: {
     title: 'Title',
@@ -19,6 +20,29 @@ export default {
   parameters: {
     ...getDocsParameters(),
   },
+  decorators: [
+    (story) => (
+      <>
+        <InlineNotification
+          className="page-layouts__banner"
+          actions={
+            <NotificationActionButton
+              href="https://carbon-for-ibm-products.netlify.app/?path=/story/ibm-products-components-cards-productivecard--default"
+              rel="noopener noreferrer"
+              target="_blank"
+            >
+              More info
+            </NotificationActionButton>
+          }
+          kind="info"
+          subtitle="This component is now legacy. Please click to see the go-forward component"
+          title=""
+          hideCloseButton
+        />
+        {story()}
+      </>
+    ),
+  ],
 };
 
 export const Default = (args) => <TrendingCard {...args} />;