+++ packages/security/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.9](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-security@2.11.8...@carbon/ibm-security@2.11.9) (2022-05-03)
+
+**Note:** Version bump only for package @carbon/ibm-security
+
+
+
+
+
+## [2.11.8](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-security@2.11.7...@carbon/ibm-security@2.11.8) (2022-04-26)
+
+**Note:** Version bump only for package @carbon/ibm-security
+
+
+
+
+
+## [2.11.7](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-security@2.11.6...@carbon/ibm-security@2.11.7) (2022-04-19)
+
+**Note:** Version bump only for package @carbon/ibm-security
+
+
+
+
+
 ## [2.11.6](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-security@2.11.5...@carbon/ibm-security@2.11.6) (2022-04-05)